fix a name error
