+++ src/constructive_methods/cubic_arrangement_method.py
@@ -17,7 +17,7 @@ class CubicArrangementMethod(ConstructiveMethod):
 
         pass
     
-    def CreatInitialPackings(self):
+    def CreateInitialPackings(self):
 
         p_id = 1
         i_row = 1
@@ -61,4 +61,4 @@ class CubicArrangementMethod(ConstructiveMethod):
                 i_thick = 1
             i_row += 1
             i_col = 1
-        print("Creat CUBIC spheres finished!\t")
+        print("Create CUBIC spheres finished!\t")
