+++ src/data_processing/pre_processing/old_version/create_particles_input_parameters.json
