+++ src/dynamic_methods/radius_expansion_method.py
@@ -12,7 +12,7 @@ __license__     = "BSD 2-Clause License"
 import os
 
 from dynamic_methods.dynamic_method import DynamicMethod
-from data_processing.pre_processing import creat_particles_inside_of_a_domain
+from data_processing.pre_processing import create_particles_inside_of_a_domain
 
 class RadiusExpansionMethod(DynamicMethod):
 
@@ -20,9 +20,9 @@ class RadiusExpansionMethod(DynamicMethod):
 
         pass
 
-    def CreatInitialCases(self):
+    def CreateInitialCases(self):
 
-        CreatIniCases = creat_particles_inside_of_a_domain.CreatParticlesInsideOfADomain()
+        CreateIniCases = create_particles_inside_of_a_domain.CreateParticlesInsideOfADomain()
         RVE_size = [self.parameters["domain_length_x"], self.parameters["domain_length_y"], self.parameters["domain_length_z"]]
         packing_num = self.parameters["packing_num"]
         domain_scale_multiplier = self.parameters["random_particle_generation_parameters"]["domain_scale_multiplier"]
@@ -30,10 +30,10 @@ class RadiusExpansionMethod(DynamicMethod):
 
         packing_cnt = 1
         while packing_cnt <= packing_num:
-            CreatIniCases.Initialize(RVE_size, domain_scale_multiplier, packing_cnt, self.ini_path)
-            CreatIniCases.CreatParticles(RVE_size)
+            CreateIniCases.Initialize(RVE_size, domain_scale_multiplier, packing_cnt, self.ini_path)
+            CreateIniCases.CreateParticles(RVE_size)
             aim_folder_name = "case_" + str(packing_cnt)
-            CreatIniCases.WriteOutGIDData(aim_folder_name, aim_file_name)
+            CreateIniCases.WriteOutGIDData(aim_folder_name, aim_file_name)
             packing_cnt += 1
 
     def RunDEM(self):
