+++ src/utilities/gravitational_deposition_method_run.py
@@ -389,7 +389,7 @@ class GravationalDepositionMethodRun(DEMAnalysisStage):
 
     def WriteOutMdpaFileOfParticles(self, output_file_name):
 
-        self.clear_old_and_creat_new_show_packing_case_folder()
+        self.clear_old_and_create_new_show_packing_case_folder()
         aim_path_and_name = os.path.join(os.getcwd(), 'show_packing', output_file_name)
 
         # clean the exsisted file first
@@ -447,7 +447,7 @@ class GravationalDepositionMethodRun(DEMAnalysisStage):
         else: # for linux
             os.system("python3 show_packing.py")
 
-    def clear_old_and_creat_new_show_packing_case_folder(self):
+    def clear_old_and_create_new_show_packing_case_folder(self):
 
         aim_path = os.path.join(os.getcwd(),'show_packing')
